refactor: clean files structure
