+++ connections/chromadb.py
@@ -1,11 +1,10 @@
-from typing import Union, Tuple, List, Set
+from typing import Union, Tuple, List
 
-from utils import log
+from utils.utils import log
 
 from langchain_community.vectorstores import Chroma as LangchainChromaVectorstore
 from langchain_openai import OpenAIEmbeddings
 import chromadb
-import tiktoken
 
 
 def get_database_client(host: str, port: str) -> chromadb.HttpClient:
