+++ data_handler/card2_embed_data.log
