+++ data_handler/card3_query_embeddings.log
